+++ .gitignore
@@ -8,6 +8,7 @@ temp
 world_cache/
 worlds
 debug_log.txt
+debug.log
 docs/bug1_bisection.md
 .manifold-intel.db
 world_generation/
