+++ src/graphical_core/compute_cull.rs
@@ -15,8 +15,6 @@ pub struct CullPushConstants {
     pub phase: u32,
     /// Opaque block mask (mesh shader path) or draw buffer offset (legacy path).
     pub draw_offset: u32,
-    /// Planet radius in blocks; used by the horizon culling early-out.
-    pub planet_radius: f32,
     /// 1 = stereo (test both eye matrices for occlusion), 0 = mono (eye 0 only).
     pub stereo: u32,
     pub _pad: [f32; 2],
@@ -162,7 +160,6 @@ mod tests {
         assert_eq!(memoffset::offset_of!(CullPushConstants, screen_size), 112);
         assert_eq!(memoffset::offset_of!(CullPushConstants, phase), 120);
         assert_eq!(memoffset::offset_of!(CullPushConstants, draw_offset), 124);
-        assert_eq!(memoffset::offset_of!(CullPushConstants, planet_radius), 128);
         assert_eq!(memoffset::offset_of!(CullPushConstants, stereo), 132);
     }
